refactor: add burger component
