+++ src/components/Header/header.module.scss
@@ -96,7 +96,6 @@
 .MuiButtonBase {
   position: relative;
   width: 100px;
-  background: #000;
   font-size: 14px;
   font-weight: bold;
   padding: 0;
