+++ src/components/Header/header.tsx
@@ -1,13 +1,11 @@
 import { useNavigate } from 'react-router-dom';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 
 import { useAppDispatch, useAppSelector } from 'redux/hooks';
 import { logOut, setOpenUserPage } from 'redux/authSlice';
 
 import { HeaderUserButtons, HeaderUserLinks } from './HeaderButtons';
-
-import MenuIcon from '@mui/icons-material/Menu';
-import { IconButton, Paper, Box, Drawer } from '@mui/material';
+import { HeaderBurger } from './HeaderBurger/headerBurger';
 
 import styles from './header.module.scss';
 
@@ -15,14 +13,8 @@ const Header = () => {
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
   const { isAuthorized, isOpenUserPage } = useAppSelector((state) => state.auth);
-  const [open, setOpen] = useState(false);
-
-  const toggleDrawer = (prop: boolean) => {
-    setOpen(prop);
-  };
 
   const goOut = () => {
-    dispatch(setOpenUserPage(false));
     localStorage.removeItem('pma_token');
     dispatch(logOut());
     navigate('/');
@@ -68,65 +60,15 @@ const Header = () => {
             )}
           </ul>
         </nav>
-        <IconButton
-          edge="start"
-          aria-label="open drawer"
-          onClick={() => toggleDrawer(!open)}
-          sx={{
-            mr: '5px',
-            display: {
-              xs: 'flex',
-              md: 'none',
-            },
-            alignItems: 'center',
-            width: '45px',
-            height: '45px',
-            padding: '1px',
+        <HeaderBurger
+          func={{
+            goSignIn,
+            goSignUp,
+            goOut,
+            goUserProfile,
+            goBoards,
           }}
-        >
-          <MenuIcon
-            sx={{
-              fill: 'blue',
-              width: '90%',
-              height: '80%',
-              padding: '3px',
-            }}
-          />
-        </IconButton>
-        <Drawer
-          anchor="right"
-          open={open}
-          onClose={() => toggleDrawer(false)}
-          sx={{ width: '500px' }}
-          keepMounted
-        >
-          <Box
-            sx={{
-              height: '100%',
-            }}
-          >
-            <Paper
-              sx={{
-                pt: '50px',
-                width: '320px',
-                height: '100%',
-              }}
-            >
-              <ul className={styles.headerNavAdaptive} onClick={() => toggleDrawer(false)}>
-                {!isAuthorized ? (
-                  <HeaderUserLinks onSignIn={goSignIn} onSignUp={goSignUp} />
-                ) : (
-                  <HeaderUserButtons
-                    openUserPage={isOpenUserPage}
-                    onClickOut={goOut}
-                    onClickUser={goUserProfile}
-                    onGoBoards={goBoards}
-                  />
-                )}
-              </ul>
-            </Paper>
-          </Box>
-        </Drawer>
+        />
       </div>
     </header>
   );
